# Conflicts:
#	src/app/reducers/templates/redux/template.reducers.ts
#	src/app/reducers/vm/redux/vm.effects.ts
#	src/app/security-group/services/security-group.service.ts
#	src/app/shared/services/affinity-group.service.ts
#	src/app/shared/services/snapshot.service.ts
#	src/app/shared/services/ssh-keypair.service.ts
#	src/app/shared/services/tags/mark-for-removal.service.ts
#	src/app/shared/services/tags/security-group-tag.service.ts
#	src/app/shared/services/tags/volume-tag.service.ts
#	src/app/shared/services/volume.service.ts
#	src/app/template/shared/iso.service.ts
#	src/app/vm/container/service-offering-dialog.container.ts
#	src/app/vm/shared/vm-entity-deletion.service.ts
#	src/app/vm/shared/vm.model.ts
#	src/app/vm/shared/vm.service.ts
#	src/app/vm/vm-creation/services/vm-creation-security-group.service.spec.ts
#	src/app/vm/vm-sidebar/vm-detail/ssh/ssh-keypair.component.ts

+++ src/app/shared/actions/volume-actions/volume-actions.container.ts
@@ -1,8 +1,10 @@
 import { Component, Input } from '@angular/core';
 import { Store } from '@ngrx/store';
+import { filter, onErrorResumeNext } from 'rxjs/operators';
+
 import { DialogService } from '../../../dialog/dialog-service/dialog.service';
-import { State } from '../../../reducers/index';
-import { Volume } from '../../models/volume.model';
+import { State } from '../../../reducers';
+import { Volume } from '../../models';
 import { AuthService } from '../../services/auth.service';
 
 import * as volumeActions from '../../../reducers/volumes/redux/volumes.actions';
@@ -35,14 +37,14 @@ export class VolumeActionsContainerComponent {
   public onVolumeDelete(volume: Volume): void {
     this.dialogService.confirm({
       message: 'DIALOG_MESSAGES.VOLUME.CONFIRM_DELETION'
-    })
-      .onErrorResumeNext()
-      .filter(res => Boolean(res))
+    }).pipe(
+      onErrorResumeNext(),
+      filter(res => Boolean(res)))
       .subscribe(() => {
-        if (!!volume.snapshots.length) {
-          this.dialogService.confirm({ message: 'DIALOG_MESSAGES.SNAPSHOT.CONFIRM_ALL_DELETION' })
-            .onErrorResumeNext()
-            .filter(res => Boolean(res))
+        if (volume.snapshots && !!volume.snapshots.length) {
+          this.dialogService.confirm({ message: 'DIALOG_MESSAGES.SNAPSHOT.CONFIRM_ALL_DELETION' }).pipe(
+            onErrorResumeNext(),
+            filter(res => Boolean(res)))
             .subscribe(() => this.store.dispatch(new snapshotActions.DeleteSnapshots(volume.snapshots)));
         }
         this.store.dispatch(new volumeActions.DeleteVolume(volume));
